emoji: Move emoji storage and search logic to a top-level module
This way we can share the logic with other components or the package consumers
